Move `IOperableEventTile` and `IEventTileOps` out of `MessageContextMenu`
